+++ src/workers/sync-engine/modules/files/application/FileCreator.ts
@@ -3,7 +3,7 @@ import { FilePath } from '../domain/FilePath';
 import { File } from '../domain/File';
 import { FileRepository } from '../domain/FileRepository';
 import { FileSize } from '../domain/FileSize';
-import { EventBus } from '../../shared/domain/WebdavServerEventBus';
+import { EventBus } from '../../shared/domain/EventBus';
 import { RemoteFileContents } from '../../contents/domain/RemoteFileContents';
 import { FileDeleter } from './FileDeleter';
 import { PlatformPathConverter } from '../../shared/test/helpers/PlatformPathConverter';
